refactor(ci-flaky): modularize config handling
